Practice and Unit Test Final

+++ Technical-Challenges/Spring-Data-JPA/src/main/java/co/com/practical/controller/EmployeeRestController.java
@@ -35,6 +35,11 @@ public class EmployeeRestController {
 
 	@GetMapping("/consultId/{id}")
 	public ResponseEntity<Employee> consultxID (@PathVariable("id") Long id) {
+		Employee employee = employeeService.getConsultId(id);
+		if(employee == null)
+		{
+			throw new RuntimeException("Employee Id not found - " + employee);
+		}
 		return new ResponseEntity<>(employeeService.getConsultId(id),HttpStatus.OK);		
 
 	}
@@ -55,6 +60,10 @@ public class EmployeeRestController {
 	@DeleteMapping("/delete/{id}")
 	public ResponseEntity<String> deleteEmployee (@PathVariable ("id") Long id) {
 		Employee employee = employeeService.getConsultId(id);
+		if(employee == null)
+		{
+			throw new RuntimeException("Employee Id not found - " + employee);
+		}
 		employeeService.delete_Employee(employee);
 		return new ResponseEntity<>("Employee delete", HttpStatus.OK);
